Planning condition tracker
Added a new tracking system for planning conditions for individual
projects.
Added a "merge companies" option to fold multiple entries for the same
company into a single one, and correct corresponding items elsewhere
(would be easier if we avoiding allowing multiple entries for the same
company in the first place).
Some general tidying of code - lots of "print" functions replaced with
"echo".